Minor bug fix related to printing of cache full amount.
git-svn-id: svn+ssh://circa.ecs.vuw.ac.nz/u/staff/djp/projects/repository/tuttepoly/trunk@1676 cab769db-1af3-0310-8bf3-f2d376eea66d

+++ tutte/cache/simple_cache.hpp
@@ -115,7 +115,7 @@ public:
   int num_buckets() { return nbuckets; }
 
   // get space used by cache in bytes
-  unsigned int size() {  return next_p - start_p; }
+  uint64_t size() {  return next_p - start_p; }
   // get available space in bytes
   uint64_t capacity() { return bufsize; }
